requirements.txt
# Core Framework
Flask==3.0.0
gunicorn==21.2.0

# LINE Bot SDK
line-bot-sdk==3.6.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Optimization (シフト最適化)
PuLP==2.7.0

# Date/Time handling
python-dateutil==2.8.2
pytz==2023.3

# Environment variables
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Error monitoring
sentry-sdk[flask]==1.38.0

# Data validation
pydantic==2.5.2

# Excel handling (Phase 2で使用予定)
openpyxl==3.1.2
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.7
